commands: move kv subcommands to subdirs
