+++ command/commands.go
@@ -18,11 +18,11 @@ import (
 	"github.com/hashicorp/consul/command/keygen"
 	"github.com/hashicorp/consul/command/keyring"
 	"github.com/hashicorp/consul/command/kv"
-	"github.com/hashicorp/consul/command/kvdel"
-	"github.com/hashicorp/consul/command/kvexp"
-	"github.com/hashicorp/consul/command/kvget"
-	"github.com/hashicorp/consul/command/kvimp"
-	"github.com/hashicorp/consul/command/kvput"
+	kvdel "github.com/hashicorp/consul/command/kv/del"
+	kvexp "github.com/hashicorp/consul/command/kv/exp"
+	kvget "github.com/hashicorp/consul/command/kv/get"
+	kvimp "github.com/hashicorp/consul/command/kv/imp"
+	kvput "github.com/hashicorp/consul/command/kv/put"
 	"github.com/hashicorp/consul/command/leave"
 	"github.com/hashicorp/consul/command/lock"
 	"github.com/hashicorp/consul/command/maint"
