+++ command/kv/del/kv_delete.go
@@ -1,4 +1,4 @@
-package kvdel
+package del
 
 import (
 	"flag"
