+++ command/kv/del/kv_delete_test.go
@@ -1,4 +1,4 @@
-package kvdel
+package del
 
 import (
 	"strconv"
