+++ command/kv/put/kv_put.go
@@ -1,4 +1,4 @@
-package kvput
+package put
 
 import (
 	"bytes"
