+++ command/kv/put/kv_put_test.go
@@ -1,4 +1,4 @@
-package kvput
+package put
 
 import (
 	"bytes"
